Allow overwriting the `delay_seconds` attibute at call time
This makes it easier to call a task immediately, and let that task call itself
after a certain delay in case it's waiting for something to happen.

The way to do this before would be to have two different tasks which have the
same functionality but with the `delay_seconds` attribute being the only
difference.

It can also be used to clear any `delay_seconds` properties on other tasks by
setting it to either `0` or `None`.