Change orphan child checking of parent id.
We have logic for the child processes that runs to ensure the
parent hasnt died and they have become orphans. This previously
checked if the parent id was 1 and then assume the orphan should
eliminate itself because the parent has gone away. This logic
does not work in a container-based world because the parent id
could very well have a process id of 1. This resulted in the
childrend immediately eliminating themselves. This fix is to store
the parent id on startup and change the logic for orhpan checking
to see if the childs parent id is different than it was on startup.